+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Setup TURN server
   ([#621](https://github.com/chatmail/relay/pull/621))
 
+- cmdeploy: make --ssh-host work with localhost
+  ([#659](https://github.com/chatmail/relay/pull/659))
+
 - Update iroh-relay to 0.35.0
   ([#650](https://github.com/chatmail/relay/pull/650))